// frontend/src/shared/api/generated/chat/models/messageResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Chat
 * OpenAPI spec version: 0.1.0
 */
import type { AttachmentResponse } from './attachmentResponse';
import type { MessageKind } from './messageKind';
import type { MessageResponseData } from './messageResponseData';

export interface MessageResponse {
  id: string;
  conversation_id: string;
  sender_id: string | null;
  kind: MessageKind;
  text: string;
  task_id: string | null;
  client_id: string | null;
  data: MessageResponseData;
  attachments: AttachmentResponse[];
  created_at: string;
}